make plugins working

+++ Packages/imap/f_mail.pas
@@ -20,7 +20,6 @@ type
     Splitter1: TSplitter;
     GroupBox2: TGroupBox;
     WebBrowser1: TWebBrowser;
-    PageProducer1: TPageProducer;
     LB: TListBox;
     Splitter2: TSplitter;
     procedure FormCreate(Sender: TObject);
@@ -63,6 +62,7 @@ type
     procedure UpdateMail( mail : TIdMessage );
 
     procedure clearFiles;
+    function getContent : string;
   public
     procedure ForceClose( force : boolean);
   end;
@@ -74,7 +74,7 @@ implementation
 
 uses
   u_TPluginImap, m_mail, System.JSON, u_json, system.DateUtils, system.IOUtils,
-  System.Types, System.StrUtils, IdText, IdAttachmentFile, u_mail_decoder;
+  System.Types, System.StrUtils, u_mail_decoder;
 
 {$R *.dfm}
 
@@ -85,6 +85,8 @@ var
   arr : TStringDynArray;
   i   : integer;
 begin
+  if not DirectoryExists(m_wwwRoot) then exit;
+
   arr := TDirectory.GetFiles(m_wwwRoot);
   for i := low(arr) to high(arr) do
       DeleteFile(arr[i]);
@@ -111,41 +113,59 @@ begin
 end;
 
 procedure TMailForm.FormCreate(Sender: TObject);
-var
-  data : TJSONObject;
-  req  : TJSONObject;
+  function convJson( obj : TJSONObject ) : TJSONObject;
+  begin
+    Result := JFromText(formatJSON(obj));
+    if Assigned(obj) then
+      obj.Free;
 
+  end;
   procedure getMailCfg;
+  var
+    data : TJSONObject;
+    req  : TJSONObject;
   begin
     req := TJSONObject.Create;
     JReplace( req, 'cmd', 'mailconfig');
-    data := PluginImap.Data.getConfigData(req);
+
+    data := convJson(PluginImap.Data.getConfigData(req));
     if Assigned(data) then begin
       MailMod.config(data);
       data.Free;
     end;
   end;
   procedure getHttpCfg;
+  var
+    data : TJSONObject;
+    req  : TJSONObject;
+    obj : TJSONObject;
   begin
     req := TJSONObject.Create;
     JReplace( req, 'cmd', 'htmlconfig');
-    data := PluginImap.Data.getConfigData(req);
+
+    obj := PluginImap.Data.getConfigData(req);
+    data := convJson(obj);
     if Assigned(data) then begin
       m_wwwRoot := JString( data, 'wwwroot');
       m_wwwPort := JInt(    data, 'port');
       data.Free;
     end;
+
     if m_wwwRoot <> '' then begin
       m_wwwRoot := TPath.Combine( m_wwwRoot, 'mail');
       ForceDirectories(m_wwwRoot);
     end;
   end;
 begin
+  ComboBox1.Text := '';
+
   PluginImap.Data.WndHandler.registerForm(self);
   MailMod := TMailMod.create(self);
   m_html  := TStringList.Create;
 
   VST.NodeDataSize := sizeof(TVSTData);
+//  m_wwwRoot := TPath.GetTempPath;
+//  ForceDirectories(m_wwwRoot));
 
   getMailCfg;
   getHttpCfg;
@@ -160,6 +180,44 @@ begin
   FreeAndNil(MailMod);
   m_html.free;
 
+end;
+
+function TMailForm.getContent: string;
+  function addrList( list : TIdEmailAddressList ): string;
+  var
+    i : integer;
+  begin
+    Result := '';
+    for i := 0 to pred(list.Count) do
+      Result := Result + prettyName(list.Items[i])+';';
+    if Result <> '' then
+      SetLength( Result, length(Result)-1);
+  end;
+  function repHTML( text : string ) : string;
+  begin
+    Result := System.StrUtils.ReplaceText(text, '<', '&lt;' );
+    Result := System.StrUtils.ReplaceText(Result, '<', '&gt;' );
+  end;
+var
+  rs   : TResourceStream;
+  list : TStringList;
+begin
+  list := TStringList.Create;
+  RS := TResourceStream.Create(HInstance, 'template', RT_RCDATA);
+  list.LoadFromStream(RS);
+  rs.Free;
+  Result := list.Text;
+  list.Free;
+
+  Result := ReplaceText(Result, '<#date>',    FormatDateTime('ddd, dd.mmmm.yyyy hh:nn', m_msg.Date ));
+  Result := ReplaceText(Result, '<#sender>',  repHTML(prettyName(m_msg.From)));
+  Result := ReplaceText(Result, '<#subject>', m_msg.Subject );
+  Result := ReplaceText(Result, '<#an>',      repHTML(addrList(m_msg.Recipients)));
+  if m_msg.CCList.Count > 0 then
+    Result := ReplaceText(Result, '<#cc>', 'cc:'+repHTML(addrList(m_msg.CCList)))
+  else
+    Result := ReplaceText(Result, '<#cc>', '' );
+  Result := ReplaceText(Result, '<#text>',    m_html.Text);
 end;
 
 procedure TMailForm.PageProducer1HTMLTag(Sender: TObject; Tag: TTag;
@@ -231,7 +289,7 @@ begin
   m_html.Text := decoder.Html.Text;
 
   fname := TPath.Combine(m_wwwRoot, 'index.html');
-  list.Text := PageProducer1.Content;
+  list.Text := getContent;
   list.SaveToFile(fname);
   WebBrowser1.Navigate(fname);
   list.Free;
